refactor(python-sdk): move timeout_to_ms to e2b/envd/utils
