+++ packages/python-sdk/e2b/envd/rpc.py
@@ -39,18 +39,6 @@ _DEFAULT_RPC_ERROR_MAP: dict[Code, Callable[[str], Exception]] = {
 }
 
 
-def timeout_to_ms(timeout: Optional[float]) -> Optional[int]:
-    """Convert a timeout in seconds to the ``timeout_ms`` connectrpc calls
-    expect. ``None`` and ``0`` (timeout disabled) map to ``None`` — connectrpc
-    treats a non-positive deadline as already expired. Positive values map to
-    at least 1 ms so a sub-millisecond timeout stays a deadline instead of
-    the 0 that connectrpc's ``timeout_ms or default`` fallback would
-    discard."""
-    if not timeout:
-        return None
-    return max(1, round(timeout * 1000))
-
-
 # pyqwest raises the builtin ConnectionError for connection-establishment
 # failures and TimeoutError for its transport timeouts (both OSError
 # subclasses); failures after the connection is up raise its ReadError /
