+++ packages/python-sdk/e2b/sandbox_async/commands/command.py
@@ -18,8 +18,8 @@ from e2b.envd.rpc import (
     authentication_header,
     extract_start_pid,
     ahandle_rpc_exception_with_health,
-    timeout_to_ms,
 )
+from e2b.envd.utils import timeout_to_ms
 from e2b.envd.client_async import as_stream, create_rpc_client, first_event
 from e2b.envd.versions import ENVD_COMMANDS_STDIN, ENVD_ENVD_CLOSE
 from e2b.exceptions import SandboxException
