+++ packages/python-sdk/e2b/sandbox_sync/commands/command.py
@@ -21,8 +21,8 @@ from e2b.envd.rpc import (
     authentication_header,
     extract_start_pid,
     handle_rpc_exception_with_health,
-    timeout_to_ms,
 )
+from e2b.envd.utils import timeout_to_ms
 from e2b.envd.client_sync import as_stream, create_rpc_client
 from e2b.envd.versions import ENVD_COMMANDS_STDIN, ENVD_ENVD_CLOSE
 from e2b.exceptions import SandboxException
