+++ packages/python-sdk/e2b/sandbox_sync/filesystem/watch_handle.py
@@ -12,8 +12,8 @@ from e2b.envd.filesystem.filesystem_pb import (
 from e2b.envd.rpc import (
     authentication_header,
     handle_rpc_exception_with_health,
-    timeout_to_ms,
 )
+from e2b.envd.utils import timeout_to_ms
 from e2b.sandbox.filesystem.filesystem import map_entry_info
 from e2b.sandbox.filesystem.watch_handle import FilesystemEvent, map_event_type
