+++ packages/python-sdk/tests/test_envd_rpc_exception.py
@@ -9,7 +9,6 @@ from e2b.envd.rpc import (
     handle_rpc_exception,
     handle_rpc_exception_with_health,
     is_transport_failure,
-    timeout_to_ms,
 )
 from e2b.exceptions import (
     AuthenticationException,
@@ -233,17 +232,3 @@ async def test_async_health_check_failure_returns_raw_error():
     original = _stream_reset()
     err = await ahandle_rpc_exception_with_health(original, check)
     assert err is original
-
-
-def test_timeout_to_ms():
-    assert timeout_to_ms(60) == 60_000
-    assert timeout_to_ms(0.5) == 500
-    assert timeout_to_ms(0.1) == 100
-    # A positive sub-millisecond timeout must stay a deadline — a 0 would be
-    # discarded by connectrpc's `timeout_ms or default` fallback, disabling
-    # the deadline entirely.
-    assert timeout_to_ms(0.0005) == 1
-    # Disabled timeouts must map to None — connectrpc treats a non-positive
-    # deadline as already expired.
-    assert timeout_to_ms(0) is None
-    assert timeout_to_ms(None) is None
